LayoutTests/compositing/overflow-trumps-transform-style-expected.txt

PASS parentNodeTransformStyle('transform-style-preserve-3d') is "preserve-3d"
PASS parentNodeTransformStyle('transform-style-flat') is "flat"
PASS parentNodeTransformStyle('overflow-hidden-with-transform-style-preserve-3d') is "preserve-3d"
PASS parentNodeTransformStyle('overflow-hidden-with-transform-style-flat') is "flat"
PASS getBoundingClientWidth('transform-style-preserve-3d') is not getBoundingClientWidth('transform-style-flat')
PASS getBoundingClientWidth('overflow-hidden-with-transform-style-flat') is getBoundingClientWidth('transform-style-flat')
PASS getBoundingClientWidth('overflow-hidden-with-transform-style-preserve-3d') is getBoundingClientWidth('transform-style-flat')
PASS successfullyParsed is true

TEST COMPLETE
Test that an element with mix-blend-mode other than normal forces the used value of transform-style to flat, but not the computed value..
